+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - none
 
 # v2023.10.1-beta
+- python - Revamp python-c-extension to use nodejs-like-api dbClose, dbExec, dbFileImportOrExport, dbNoop, dbOpen.
 - sqlmath - Remove unused sql-function-prm in sinefit_extra(), stp.
 - sqlmath - Add sql-function-prm in sinefit_extra(), predict_cnr, predict_cos, predict_sin.